added alcohol addiction, reorganized some files
