+++ Resources/Locale/en-US/_NF/addiction/addictions.ftl
@@ -1,0 +1,38 @@
+# Generic symptoms shared by addictions
+addiction-symptom-nausea-low-0 = You tummy doesn't feel good.
+addiction-symptom-nausea-low-1 = You stomach is a little upset.
+addiction-symptom-nausea-high-0 = You're on the verge of vomiting.
+addiction-symptom-nausea-high-1 = You throw-up a little in your mouth.
+
+addiction-symptom-headache-low-0 = This headache is terrible.
+addiction-symptom-headache-low-1 = You have a headache.
+addiction-symptom-headache-low-2 = This headache isn't going away.
+addiction-symptom-headache-high-0 = Your headache pounds, and pounds.
+addiction-symptom-headache-high-1 = This migraine is the worst.
+addiction-symptom-headache-high-2 = Need something for this pounding headache.
+
+addiction-stimulant-crave-low-0 = You could really use a pick me up
+addiction-stimulant-crave-low-1 = You're tired, you deserve a strong boost
+addiction-stimulant-crave-low-2 = {$lastReagent} felt really good, you should get some more
+addiction-stimulant-crave-low-3 = Things were more exciting with {$lastReagent}
+addiction-stimulant-crave-low-4 = Things were better with {$lastReagent}
+# depression is a common symptom
+addiction-stiumlant-crave-high-0 = This sucks. Where can you find some {$lastReagent}?
+addiction-stimulant-crave-high-1 = Life is such a downer. You NEED a hit of {$lastReagent} NOW!
+addiction-stimulant-crave-high-2 = Nothing really matters, except that next dose of {$lastReagent}
+addiction-stimulant-crave-high-3 = YOU NEED { CAPITALIZE($lastReagent) }, NOW!
+addiction-stimulant-crave-high-4 = {$lastReagent}, {$lastReagent}, {$lastReagent} where is it?!
+
+addiction-alcohol-crave-low-0 = Need to take that edge off.
+addiction-alcohol-crave-low-1 = Things were smoother with a drink.
+addiction-alcohol-crave-low-2 = Sobriety is overrated.
+addiction-alcohol-crave-low-3 = When can you get another drink?
+addiction-alcohol-crave-low-4 = Alcohol sounds good right about now.
+
+addiction-alcohol-crave-high-0 = People are annoying. Much more tolerable intoxicated.
+addiction-alcohol-crave-high-1 = This headache is horrible. A drink will help.
+addiction-alcohol-crave-high-2 = How long until another drink?
+addiction-alcohol-crave-high-3 = You always feel sick when sober.
+addiction-alcohol-crave-high-4 = Where is that sweet alcohol? You could use it now.
+addiction-alcohol-crave-high-5 = Nothing really fixes this headache but alcohol.
+
