--- Resources/Locale/en-US/_NF/addiction/stimulant.ftl
@@ -1,11 +1,0 @@
-addiction-stimulant-crave-low-0 = You could really use a pick me up
-addiction-stimulant-crave-low-1 = You're tired, you deserve a strong boost
-addiction-stimulant-crave-low-2 = {$lastReagent} felt really good, you should get some more
-addiction-stimulant-crave-low-3 = Things were more exciting with {$lastReagent}
-addiction-stimulant-crave-low-4 = Things were better with {$lastReagent}
-# depression is a common symptom
-addiction-stiumlant-crave-high-0 = This sucks. Where can you find some {$lastReagent}?
-addiction-stimulant-crave-high-1 = Life is such a downer. You NEED a hit of {$lastReagent} NOW!
-addiction-stimulant-crave-high-2 = Nothing really matters, except that next dose of {$lastReagent}
-addiction-stimulant-crave-high-3 = YOU NEED { CAPITALIZE($lastReagent) }, NOW!
-addiction-stimulant-crave-high-4 = {$lastReagent}, {$lastReagent}, {$lastReagent} where is it?!
